requirement.txt

anyjson==0.3.3
celery==3.1.19
click==6.6
django-celery==3.1.17
django-cors-headers==1.1.0
django-filter==0.13.0
django-model-utils==2.3.1
django-simple-captcha==0.5.2
djangorestframework==3.2.3
djangorestframework-jwt==1.8.0
Jinja2==2.8
kombu==3.0.30
livereload==2.4.1
Markdown==2.6.6
MarkupSafe==0.23
mkdocs==0.15.3
mkdocs-bootstrap==0.1.1
mkdocs-bootswatch==0.4.0
mysqlclient==1.3.6
Pillow==2.9.0
PyJWT==1.4.0
pypng==0.0.18
pytz==2015.7
PyYAML==3.11
requests==2.11.1
six==1.9.0
wheel==0.24.0
